project/ml-api/requirements.txt
Flask==3.0.3
flask-restx==1.3.0
flask-cors==4.0.1
python-dotenv==1.0.1
joblib==1.4.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
seaborn==0.13.2
